pyproject.toml
[tool.poetry]
name = "objectherkenning_openbare_ruimte"
version = "0.1.0"
description = "Objectherkenning Openbare Ruimte is a project about recognising objects from public space images."
authors = [
    "Sebastian Davrieux <s.davrieux@amsterdam.nl>",
    "Diana Epureanu <d.epureanu@amsterdam.nl>",
    "Andrea Lombardo <a.lombardo@amsterdam.nl>",
    "Daan Bloembergen <d.bloembergen@amsterdam.nl>"
]

[tool.poetry.dependencies]
python = "~3.8"
pre-commit = {version = "^3.2.0", source = "PyPI"}
aml_interface = { git = "https://github.com/Computer-Vision-Team-Amsterdam/AML-Interface.git",  tag = "v1.0.0" }
cvtoolkit = { git = "https://github.com/Computer-Vision-Team-Amsterdam/CVToolkit.git",  tag = "v1.0.2" }
pydantic = {version = "^1.8.2", source = "PyPI"}
azure-iot-device = {version = "^2.13.0", source = "PyPI"}
azure-ai-ml = {version = "^1.15.0", source = "PyPI"}
azure-storage-blob = {version = "^12.19.0", source = "PyPI"}
geopandas = "~0.13.2"

[tool.poetry.dev-dependencies]
pytest = {version = "^7.0.1", source = "PyPI" }
pytest-cov = {version = "^4.0.0", source = "PyPI" }
ipykernel = "^6.29.4"
ipympl = "^0.9.3"
matplotlib = "~3.7.0"

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
